Show fix of CustomWebFilterTest
As org.springframework.web.server.adapter.DefaultServerWebExchange and its sub-class org.springframework.mock.web.server.MockServerWebExchange do not define hashcode() implementations, your webFilterChain.filter(mutatedServerWebExchange) could not be successful, because the principle of verifying the correct mocking of a method depends on hashcode(). If this is not defined the one of the java.lang.Object gets into action and this means referential equality, which is, of course, not the case in your test. So we have to use a cool Mockito feature called argument captor. Try to figure out how it works and read the docs here: http://static.javadoc.io/org.mockito/mockito-core/2.13.0/org/mockito/Mockito.html#15